Reorganized test in it's own file
